+++ packages/card/src/__tests__/shared.js
@@ -1,9 +1,10 @@
 import React from "react";
+import { Text } from "react-native";
 import { shallow } from "enzyme";
 import Card from "../card";
 import Loading from "../card-loading";
 
-const cardProps = {
+const props = {
   image: {
     uri:
       "https://www.thetimes.co.uk/imageserver/image/methode%2Fsundaytimes%2Fprod%2Fweb%2Fbin%2F9242e576-4dfc-11e7-a20e-a11097d3353d.jpg?crop=1463%2C975%2C293%2C12"
@@ -14,73 +15,148 @@ const cardProps = {
 };
 
 export default () => {
-  it("should render a card", () => {
+  it("should render default layout", () => {
     const component = shallow(
-      <Card {...cardProps}>
-        <span>A card</span>
+      <Card {...props}>
+        <Text>A card</Text>
       </Card>
     );
 
-    expect(component).toMatchSnapshot("Render a Card");
+    expect(component).toMatchSnapshot("1. Renders default layout");
   });
 
-  it("should render the loading state", () => {
+  it("should render without an image when null", () => {
     const component = shallow(
-      <Card {...cardProps} isLoading>
-        <span>Loading state</span>
+      <Card {...props} image={null}>
+        <Text>No image</Text>
       </Card>
     );
 
-    expect(component).toMatchSnapshot("Render the loading state");
-  });
-
-  it("should render without an image", () => {
-    const noImageProps = Object.assign({}, cardProps, {
-      image: null
-    });
-    const component = shallow(
-      <Card {...noImageProps}>
-        <span>No image</span>
-      </Card>
-    );
-
-    expect(component).toMatchSnapshot("Renders without an image");
+    expect(component).toMatchSnapshot("2. Renders without an image (null)");
   });
 
   it("should render without an image when showImage is false", () => {
-    const noImageProps = Object.assign({}, cardProps, {
-      showImage: false
-    });
     const component = shallow(
-      <Card {...noImageProps}>
-        <span>No image</span>
+      <Card {...props} showImage={false}>
+        <Text>No image</Text>
       </Card>
     );
 
     expect(component).toMatchSnapshot(
-      "Renders without an image when showImage is false"
+      "3. Renders without an image (showImage is false)"
     );
   });
 
-  it("should render without image url", () => {
-    const noImageProps = Object.assign({}, cardProps, {
-      image: {
-        uri: null
-      }
-    });
+  it("should render without an image when url is null", () => {
     const component = shallow(
-      <Card {...noImageProps}>
-        <span>No URL</span>
+      <Card {...props} image={{ uri: null }}>
+        <Text>No URL</Text>
       </Card>
     );
 
-    expect(component).toMatchSnapshot("Renders without an image URL");
+    expect(component).toMatchSnapshot(
+      "4. Renders without an image (uri is null)"
+    );
+  });
+
+  it("should render with reversed layout", () => {
+    const component = shallow(
+      <Card {...props} isReversed>
+        <Text>A card</Text>
+      </Card>
+    );
+
+    expect(component).toMatchSnapshot("5. Renders with a reversed layout");
+  });
+
+  it("should render with reversed layout and no image", () => {
+    const component = shallow(
+      <Card {...props} isReversed showImage={false}>
+        <Text>A card</Text>
+      </Card>
+    );
+
+    expect(component).toMatchSnapshot(
+      "6. Renders with a reversed layout with no image"
+    );
+  });
+
+  it("should render a loading state", () => {
+    const component = shallow(
+      <Card {...props} isLoading>
+        <Text>Loading state</Text>
+      </Card>
+    );
+
+    expect(component).toMatchSnapshot("7. Renders a loading state");
+  });
+
+  it("should render a loading component", () => {
+    const component = shallow(
+      <Loading {...props}>
+        <Text>A card</Text>
+      </Loading>
+    );
+
+    expect(component).toMatchSnapshot("8. Renders a loading component");
+  });
+
+  it("should render a loading card with no image", () => {
+    const component = shallow(
+      <Loading {...props} showImage={false}>
+        <Text>A card</Text>
+      </Loading>
+    );
+
+    expect(component).toMatchSnapshot(
+      "9. Render a loading component with no image"
+    );
+  });
+
+  it("should render a reversed loading component", () => {
+    const component = shallow(
+      <Loading {...props} isReversed>
+        <Text>A card</Text>
+      </Loading>
+    );
+
+    expect(component).toMatchSnapshot(
+      "10. Render a reversed loading component"
+    );
+  });
+
+  it("should render a reversed loading component with no image", () => {
+    const component = shallow(
+      <Loading {...props} isReversed showImage={false}>
+        <Text>A card</Text>
+      </Loading>
+    );
+
+    expect(component).toMatchSnapshot(
+      "11. Render a reversed loading component with no image"
+    );
+  });
+
+  it("should not re-render when imageRatio prop is changed", () => {
+    const component = shallow(
+      <Card {...props}>
+        <Text>Do not re-render me</Text>
+      </Card>
+    );
+
+    expect(component).toMatchSnapshot();
+
+    component.setProps({
+      imageRatio: 16 / 9
+    });
+
+    expect(component).toMatchSnapshot();
   });
 
   it("should re-render when image uri changes", () => {
     const component = shallow(
-      <Card {...cardProps}>
-        <span>Some text</span>
+      <Card {...props}>
+        <Text>Some text</Text>
       </Card>
     );
 
@@ -94,16 +170,23 @@ export default () => {
       component
         .find("TimesImage")
         .at(0)
-        .props().uri
-    ).toEqual(`${testUri}&resize=${cardProps.imageSize}`);
+        .prop("uri")
+    ).toEqual(`${testUri}&resize=${props.imageSize}`);
   });
 
   it("should re-render when image size changes", () => {
     const component = shallow(
-      <Card {...cardProps}>
-        <span>Some content</span>
+      <Card {...props}>
+        <Text>Some content</Text>
       </Card>
     );
+
+    expect(
+      component
+        .find("TimesImage")
+        .at(0)
+        .prop("uri")
+    ).toEqual(`${props.image.uri}&resize=${props.imageSize}`);
 
     component.setProps({
       imageSize: null
@@ -113,79 +196,23 @@ export default () => {
       component
         .find("TimesImage")
         .at(0)
-        .props().uri
-    ).toEqual(`${cardProps.image.uri}`);
+        .prop("uri")
+    ).toEqual(`${props.image.uri}`);
   });
 
-  it("should re-render after showing the loading state", () => {
+  it("should re-render when loading state changes", () => {
     const component = shallow(
-      <Card {...cardProps} isLoading>
-        <span>Re-render me</span>
+      <Card {...props} isLoading>
+        <Text>Re-render me</Text>
       </Card>
     );
+
+    expect(component.name()).toEqual("Loading");
 
     component.setProps({
       isLoading: false
     });
 
-    expect(component).toMatchSnapshot(
-      "Re-renders after showing the loading state"
-    );
-  });
-
-  it("should not re-render when image ratio changes", () => {
-    const component = shallow(
-      <Card {...cardProps}>
-        <span>Do not re-render me</span>
-      </Card>
-    );
-
-    component.setProps({
-      imageRatio: 2
-    });
-
-    expect(component).toMatchSnapshot(
-      "Renders Card normally and does not re-render"
-    );
-  });
-
-  it("should render a card with reversed layout", () => {
-    const component = shallow(
-      <Card {...cardProps} isReversed>
-        <span>A card</span>
-      </Card>
-    );
-
-    expect(component).toMatchSnapshot("Renders a Card with a reversed layout");
-  });
-
-  it("should render a loading card", () => {
-    const component = shallow(
-      <Loading {...cardProps}>
-        <span>A card</span>
-      </Loading>
-    );
-
-    expect(component).toMatchSnapshot("Render a loading Card");
-  });
-
-  it("should render a loading card with no image", () => {
-    const component = shallow(
-      <Loading {...cardProps} showImage={false}>
-        <span>A card</span>
-      </Loading>
-    );
-
-    expect(component).toMatchSnapshot("Render a loading Card");
-  });
-
-  it("should render a reversed loading card", () => {
-    const component = shallow(
-      <Loading {...cardProps} isReversed>
-        <span>A card</span>
-      </Loading>
-    );
-
-    expect(component).toMatchSnapshot("Render a loading Card");
+    expect(component.name()).toEqual("FadeIn");
   });
 };